+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v2.0.0
+
+- Context now shows correctly everywhere
+- It is now possible to add arguments to your HTML elements. To do this, add args to your 
+action like this `action: (args: any) => { console.log('Deleting ' + args.name); }` and
+add the arguments to your HTML element like this
+`[ctxArgs]="{id: person.id, name: person.name, message: person.message}"`.
+
+### BREAKING CHANGES
+
+- The way to add context to an HTML element has changed. From now on you need to use
+`[ctx]="'context-name'"`. This change is to fix a bug where not all menu items would show.
+
 ## v1.0.0
 
 - Add better styling